Cleaning up, moving code to source dir
